+++ libs/cardano-ledger-pretty/src/Cardano/Ledger/Pretty.hs
@@ -57,7 +57,7 @@ import Cardano.Ledger.BaseTypes (
   txIxToInt,
  )
 import Cardano.Ledger.Block (Block (..))
-import Cardano.Ledger.CertState (VState (..))
+import Cardano.Ledger.CertState (DRepState, VState (..))
 import Cardano.Ledger.Coin (Coin (..), DeltaCoin (..))
 import Cardano.Ledger.Compactible (Compactible (..))
 import Cardano.Ledger.Core
@@ -1683,6 +1683,9 @@ instance PrettyA (GenDelegPair c) where
 
 instance PrettyA (GenDelegs c) where
   prettyA = ppGenDelegs
+
+instance PrettyA (DRepState c) where
+  prettyA = viaShow
 
 instance PrettyA (VState era) where
   prettyA (VState vsDReps vsCommitteeHotKeys) =